Remove old project manager, move functionality into project/file.js
